made separate folder for components
