+++ 06-usepopcorn-app/src/components/ErrorMessage.js
@@ -1,0 +1,3 @@
+export default function ErrorMessage({ message }) {
+  return <p className="error">❌{message}</p>;
+}
