+++ 06-usepopcorn-app/src/components/Main.js
@@ -1,0 +1,3 @@
+export default function Main({ children }) {
+  return <main className="main">{children}</main>;
+}
