+++ 06-usepopcorn-app/src/components/MoviesList.js
@@ -1,0 +1,14 @@
+import MovieItem from "./MovieItem";
+export default function MoviesList({ movies, onSelectMovie }) {
+  return (
+    <ul className="list list-movies">
+      {movies?.map((movie) => (
+        <MovieItem
+          movie={movie}
+          key={movie.imdbID}
+          onSelectMovie={onSelectMovie}
+        />
+      ))}
+    </ul>
+  );
+}
