+++ 06-usepopcorn-app/src/components/WatchedMovieList.js
@@ -1,0 +1,10 @@
+import WatchedMovie from "./WatchedMovie";
+export default function WatchedMoviesList({ watched }) {
+  return (
+    <ul className="list">
+      {watched.map((movie) => (
+        <WatchedMovie movie={movie} key={movie.imdbID} />
+      ))}
+    </ul>
+  );
+}
